% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_18_helpers.R
\name{are_adjacent}
\alias{are_adjacent}
\title{Check if two vertices are adjacent in PAG}
\usage{
are_adjacent(pag, v1, v2)
}
\arguments{
\item{pag}{Adjacency matrix}

\item{v1}{First vertex}

\item{v2}{Second vertex}
}
\value{
TRUE if v1 and v2 are adjacent, FALSE otherwise
}
\description{
Check if two vertices are adjacent in PAG
}
