% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_pag.R
\name{complete_pag}
\alias{complete_pag}
\title{Complete expert knowledge}
\usage{
complete_pag(pag, rules = NULL, verbose = FALSE, debug = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix of ancestral partial mixed graph.
This is the essential graph with some circle edges oriented as per expert knowledge.
The matrix follows the convention in pcalg: 0 refers to no edge. 1 refers to circle edge mark.
2 refers to an arrowhead and 3 refers to a tail. For example, \code{pag[a, b] = 2} and \code{pag[b, a] = 1}
says that the edge is \verb{a o-> b}.}

\item{rules}{A boolean vector of length 18 indicating which rules are to be invoked}

\item{verbose}{Boolean indicating verboseness}

\item{debug}{Boolean indicating verboseness level higher than verbose}
}
\value{
A list with two named elements, \code{pag} and \code{counter}.
\code{pag} is the adjacency matrix obtained after completing all the rules turned on in the \code{rules} parameter.
\code{counter} is a vector of length 18 indicating how many times each rule was invoked.
}
\description{
Perform orientation on an ancestral partial mixed graph to restrict equivalence class to expert knowledge.
}
\seealso{
\code{\link[=generate_mag_pag]{generate_mag_pag()}} for format of adjacency matrices
}
