% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_14_helpers.R
\name{find_all_prior_to}
\alias{find_all_prior_to}
\title{Find all vertices that are prior to B relative to X}
\usage{
find_all_prior_to(pag, B, X, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix}

\item{B}{Target vertex}

\item{X}{Reference vertex}

\item{verbose}{Print debug information}
}
\value{
Vector of vertex indices that are prior to B
}
\description{
Uses BFS to build the transitive closure of the prior-to relationship
}
