% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_14_helpers.R
\name{find_circle_paths_to_targets}
\alias{find_circle_paths_to_targets}
\title{Find circle-circle paths from F1 to vertices that can reach C2}
\usage{
find_circle_paths_to_targets(pag, F1, C2, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix}

\item{F1}{Starting vertex}

\item{C2}{Target vertex (looking for Fk where Fk o-> C2)}

\item{verbose}{Print debug information}
}
\value{
List of paths (each path is a vector of vertices)
}
\description{
Finds paths where all edges are o-o and the last vertex has o-> C2
}
