% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_18_helpers.R
\name{find_possible_directed_paths}
\alias{find_possible_directed_paths}
\title{Find minimal possible directed paths from source to target}
\usage{
find_possible_directed_paths(pag, source, target, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix}

\item{source}{Source vertex (D)}

\item{target}{Target vertex (A)}

\item{verbose}{Print debug information}
}
\value{
List of paths, each path is a vector of vertices
}
\description{
A "possible directed path" is a path where edges could be oriented
to form a directed path. We look for paths following edges that
are not blocked by arrowheads pointing away from the path direction.
}
