% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_14_helpers.R
\name{find_prior_condition_3}
\alias{find_prior_condition_3}
\title{Find vertices satisfying Condition 3: Unbridged path in induced subgraph}
\usage{
find_prior_condition_3(pag, target, X, candidate_set, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix}

\item{target}{Target vertex (Fi+1)}

\item{X}{Reference vertex}

\item{candidate_set}{Candidates for Fi}

\item{verbose}{Print debug information}
}
\value{
Vector of vertices satisfying condition 3
}
\description{
This is the most complex condition - for now, return empty set
(Can be implemented if needed, but may not be critical for many cases)
}
