% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_14_helpers.R
\name{find_unbridged_paths}
\alias{find_unbridged_paths}
\title{Find unbridged paths relative to {T1, T2}}
\usage{
find_unbridged_paths(pag, T1, T2, X, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix (PAG)}

\item{T1}{First vertex in reference set}

\item{T2}{Second vertex in reference set}

\item{X}{Reference vertex (must have X o-* Fi for all Fi)}

\item{verbose}{Print debug information}
}
\value{
List of unbridged paths (each path is a vector of vertex indices)
}
\description{
An unbridged path F1 o-o ... o-o Fk relative to {T1, T2} satisfies:
\itemize{
\item There exist C1, C2 in {T1, T2}
\item Structure: C1 <-o F1 o-o F2 o-o ... o-o Fk o-> C2
(i.e., F1 o-> C1 and Fk o-> C2)
\item NO edge C1 <-> C2 exists
\item X o-* Fi for all Fi in the path
}
}
