% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_18_helpers.R
\name{get_path_second_vertex}
\alias{get_path_second_vertex}
\title{Find the first vertex after source in a path}
\usage{
get_path_second_vertex(path)
}
\arguments{
\item{path}{Vector of vertices in path}
}
\value{
Second vertex in path (or NULL if path too short)
}
\description{
For path <D, T, ..., A>, extract T
}
