% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_14_helpers.R
\name{is_prior_to}
\alias{is_prior_to}
\title{Check if vertex A is prior to B relative to X}
\usage{
is_prior_to(pag, A, B, X, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix}

\item{A}{Starting vertex}

\item{B}{Target vertex}

\item{X}{Reference vertex}

\item{verbose}{Print debug information}
}
\value{
TRUE if A is prior to B, FALSE otherwise
}
\description{
Implements Definition 3: A is prior to B relative to X if there exists
a sequence F0 = A, F1, ..., Ft = B where Fi *-o X and one of three
conditions holds for each step.
}
