% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_14_helpers.R
\name{is_prior_to_order_1}
\alias{is_prior_to_order_1}
\title{Check if any vertex in candidate_set is prior to target with order 1}
\usage{
is_prior_to_order_1(pag, target, X, candidate_set, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix}

\item{target}{Target vertex (the Fi+1 in the definition)}

\item{X}{Reference vertex}

\item{candidate_set}{Set of candidate vertices to check}

\item{verbose}{Print debug information}
}
\value{
Vector of vertices that are prior to target
}
\description{
Implements the three conditions from Definition 3
}
