% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_14.R
\name{rule.14}
\alias{rule.14}
\title{Apply Rule 14 orientation rule}
\usage{
rule.14(pag, unfVect = NULL, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix (PAG)}

\item{unfVect}{Vector of ambiguous triples (optional, not used)}

\item{verbose}{Print debug information}
}
\value{
Updated PAG with Rule 14 orientations applied
}
\description{
Rule 14: Orient X o-> B as X --> B when there exist T1, T2 such that:
\itemize{
\item Both T1 \emph{-o X o-} T2 (circle at X)
\item Both T1 and T2 are prior to B relative to X
\item EITHER:
(a) There exists an unbridged path F1 o-o ... o-o Fk relative to {T1, T2}
where X o-* Fi for all Fi, OR
(b) T1 and T2 are NOT adjacent
}
}
