% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_18.R
\name{rule.18}
\alias{rule.18}
\title{Apply Rule 18 orientation rule}
\usage{
rule.18(pag, unfVect = NULL, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix (will be modified in place)}

\item{verbose}{Print debug information}
}
\value{
Updated PAG with Rule 18 orientations applied
}
\description{
Rule 18: Orient A o-> B as A <-> B when:
\itemize{
\item A o-> B <-> D (A has circle-arrow to B, B has bidirected to D)
\item There exist two minimal possible directed paths <D, T1, ..., A> and <D, T2, ..., A>
\item Either:
(a) There is an unbridged path F1 o-o ... o-o Fk relative to {T1, T2}
such that D o-* Fi for all Fi, OR
(b) T1 and T2 are not adjacent
}
}
