% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_conjecture.R
\name{verify_conjecture}
\alias{verify_conjecture}
\title{Verify conjecture}
\usage{
verify_conjecture(pag, G, stop_first = FALSE, verbose = FALSE)
}
\arguments{
\item{pag}{Adjacency matrix of ancestral partial mixed graph. Typically, the output of \code{expertOrientR::complete_pag}}

\item{G}{Essential ancestral graph}

\item{stop_first}{Whether we should stop our search if we find some MAG that shares the same invariant edges as \code{pag}.
User should always set this to \code{FALSE} for verifying conjecture. It is set to \code{TRUE} by internal functions.}

\item{verbose}{Boolean describing verboseness level}
}
\value{
Boolean. \code{TRUE} if \code{pag} is a proper restriction. \code{FALSE} if \code{pag} is not a proper restriction.
}
\description{
Verify whether the restricted equivalence class, represented by an adjacency matrix,
is a proper restriction with respect to some admissible and consistent expert knowledge.
}
\seealso{
\code{\link[=generate_mag_pag]{generate_mag_pag()}} for format of adjacency matrices
}
